Added support for scanning/not scanning.
This entailed adding a segmented switch to the main view controller.

I also made the main view controller the "holder of the driver," because I didn't want an awkward relationship between the app delegate and the view controller.

I added a couple of delegate calls to each delegate to respond when devices are removed. However, I have not yet added the removel code to the driver.