There's still some VERY strange stuff going on with my WatchKit framework test harness target. I'm actually beginning to wonder if there's an Xcode bug.
The debugger does not seem to be matching the source with the running app.
The app does not seem to be mapping the table controllers to their proper types.
Looks like there may be some kindo fo RTTI-type issue going on.
I will continue to see if I can figure this out...